+++ src/engine/H2DE_chrono_manager.cpp
@@ -17,6 +17,7 @@ H2DE_ChronoManager::~H2DE_ChronoManager() {
 void H2DE_ChronoManager::destroyChrono(H2DE_Chrono* chrono) {
     auto it = std::find(chronos.begin(), chronos.end(), chrono);
     if (it != chronos.end()) {
+        delete *it;
         chronos.erase(it);
     }
 }
@@ -36,12 +37,12 @@ H2DE_Chrono* H2DE_ChronoManager::create(const H2DE_Time& start, bool increasing,
 }
 
 // GETTER
-H2DE_Time H2DE_ChronoManager::elapsedToTime(float elapsed) {
+H2DE_Time H2DE_ChronoManager::elapsedToTime(float elapsed) noexcept {
     H2DE_Time res = H2DE_Time();
 
-    res.milliseconds = std::clamp(H2DE::round((elapsed - H2DE::floor(elapsed)) * 1000.0f), 0, 999);
-    res.seconds = std::clamp(static_cast<int>(elapsed) % 60, 0, 59);
-    res.minutes = std::clamp((static_cast<int>(elapsed) % 3600) / 60, 0, 59);
+    res.milliseconds = H2DE::clamp(H2DE::round((elapsed - H2DE::floor(elapsed)) * 1000.0f), 0, 999);
+    res.seconds = H2DE::clamp(static_cast<int>(elapsed) % 60, 0, 59);
+    res.minutes = H2DE::clamp((static_cast<int>(elapsed) % 3600) / 60, 0, 59);
     res.hours = static_cast<int>(elapsed) / 3600;
 
     return res;