+++ src/objects/H2DE_timer_object.cpp
@@ -5,34 +5,16 @@
 H2DE_TimerObject::H2DE_TimerObject(H2DE_Engine* e, const H2DE_ObjectData& od, const H2DE_TimerObjectData& tod) : H2DE_Object(e, od), timerObjectData(tod) {
     refreshSurfaceBuffers();
     refreshMaxRadius();
-    initElapsedTime();
-    initTimeline();
+    initChrono();
 }
 
-void H2DE_TimerObject::initElapsedTime() noexcept {
-    float ms = timerObjectData.time.milliseconds / 1000.0f;
-    float sec = timerObjectData.time.seconds;
-    float min = timerObjectData.time.minutes * 60.0f;
-    float hrs = timerObjectData.time.hours * 3600.0f;
-    elapsed = ms + sec + min + hrs;
-}
-
-void H2DE_TimerObject::initTimeline() {
-    timeline = engine->createTimeline(0, H2DE_EASING_LINEAR, nullptr, [this]() {
-        int deltaTimeMultiplier = (timerObjectData.increasing) ? 1 : -1;
-        
-        elapsed += engine->getDeltaTime() * deltaTimeMultiplier;
-        elapsed = std::max(elapsed, 0.0f);
-        timerObjectData.time = H2DE_Time::toTime(elapsed);
-
-        textObject->setText(getStringifiedTime());
-
-    }, H2DE_INFINITE_LOOP, timerObjectData.pauseSensitive);
+void H2DE_TimerObject::initChrono() {
+    chrono = engine->createChrono(timerObjectData.time, timerObjectData.increasing, timerObjectData.pauseSensitive);
 }
 
 // CLEANUP
 H2DE_TimerObject::~H2DE_TimerObject() {
-    destroyTimeline();
+    destroyChrono();
 
     if (textObject != nullptr) {
         if (engine->destroyObject(textObject)) {
@@ -43,31 +25,14 @@ H2DE_TimerObject::~H2DE_TimerObject() {
     H2DE_Object::destroySurfaces(surfaces);
 }
 
-void H2DE_TimerObject::destroyTimeline() {
-    timeline->stop(false);
+void H2DE_TimerObject::destroyChrono() {
+    chrono->stop();
 }
 
 // UPDATE
 void H2DE_TimerObject::update() {
-    float epsilon = 0.5f / engine->getFPS();
-
-    for (auto it = onReachEvents.begin(); it != onReachEvents.end(); ) {
-        float elapsedDifferenceWithTarget = H2DE::abs(it->target.toElapsed() - elapsed);
-
-        if (elapsedDifferenceWithTarget <= epsilon) {
-            it->callback();
-
-            if (it->once) {
-                it = onReachEvents.erase(it);
-
-            } else {
-                ++it;
-            }
-
-        } else {
-            ++it;
-        }
-    }
+    timerObjectData.time = chrono->getTime();
+    textObject->setText(getStringifiedTime());
 }
 
 // ACTIONS
@@ -103,36 +68,23 @@ void H2DE_TimerObject::refreshMaxRadius() {
     float maxHitboxesRadius = getMaxHitboxRadius();
     float maxSurfaceRadius = getMaxSurfaceRadius(surfaces);
     
-    maxRadius = std::max(maxHitboxesRadius, maxSurfaceRadius);
-}
-
-void H2DE_TimerObject::refreshTimeline() {
-    destroyTimeline();
-    initTimeline();
+    maxRadius = H2DE::max(maxHitboxesRadius, maxSurfaceRadius);
 }
 
 void H2DE_TimerObject::onReach(const H2DE_Time& target, const std::function<void()>& callback, bool once) {
-    if (!callback) {
-        return;
-    }
-
-    H2DE_TimerObject::H2DE_OnReachEvent onReachEvent = H2DE_TimerObject::H2DE_OnReachEvent();
-    onReachEvent.target = target;
-    onReachEvent.callback = callback;
-    onReachEvent.once = once;
-    onReachEvents.push_back(onReachEvent);
+    chrono->onReach(target, callback, once);
 }
 
 void H2DE_TimerObject::pause() {
-    timeline->pause();
+    chrono->pause();
 }
 
 void H2DE_TimerObject::resume() {
-    timeline->resume();
+    chrono->resume();
 }
 
 void H2DE_TimerObject::togglePause() {
-    timeline->togglePause();
+    chrono->togglePause();
 }
 
 // GETTER
@@ -166,7 +118,7 @@ const std::string H2DE_TimerObject::getStringifiedTime() const {
 }
 
 bool H2DE_TimerObject::isPaused() const {
-    return timeline->isPaused();
+    return chrono->isPaused();
 }
 
 // SETTER
@@ -174,37 +126,42 @@ bool H2DE_TimerObject::isPaused() const {
 // -- non lerp
 void H2DE_TimerObject::setTime(const H2DE_Time& time) {
     timerObjectData.time = time;
+    chrono->setTime(time);
     refreshSurfaceBuffers();
 }
 
 void H2DE_TimerObject::setHours(uint8_t hours) {
     timerObjectData.time.hours = hours;
+    chrono->setHours(hours);
     refreshSurfaceBuffers();
 }
 
 void H2DE_TimerObject::setMinutes(uint8_t minutes) {
     timerObjectData.time.minutes = minutes;
+    chrono->setMinutes(minutes);
     refreshSurfaceBuffers();
 }
 
 void H2DE_TimerObject::setSeconds(uint8_t seconds) {
     timerObjectData.time.seconds = seconds;
+    chrono->setSeconds(seconds);
     refreshSurfaceBuffers();
 }
 
 void H2DE_TimerObject::setMilliseconds(uint16_t milliseconds) {
     timerObjectData.time.milliseconds = milliseconds;
+    chrono->setMilliseconds(milliseconds);
     refreshSurfaceBuffers();
 }
 
 void H2DE_TimerObject::setIncreasing(bool increasing) {
     timerObjectData.increasing = increasing;
-    refreshTimeline();
+    chrono->setSetIncreasing(increasing);
 }
 
 void H2DE_TimerObject::setPauseSensitive(bool pauseSensitive) {
     timerObjectData.pauseSensitive = pauseSensitive;
-    refreshTimeline();
+    chrono->setPauseSensitive(pauseSensitive);
 }
 
 // lerp